+++ ecommerce_project/myapp/serializers/__init__.py
@@ -1,1 +1,2 @@
 from .product_serializers import *
+from .review_serializers import *
