+++ ecommerce_project/myapp/views/product_related_views.py
@@ -1,9 +1,10 @@
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from ecommerce_project.myapp.models import Product, Review, Order, OrderLine, Cart, CartLine
+from ecommerce_project.myapp.models import Review, Order, OrderLine, Cart, CartLine
+from ecommerce_project.myapp.serializers import ReviewOutputSerializer, ReviewInputSerializer
 from ecommerce_project.myapp.serializers.product_related_seralizers import  \
-     ReviewOutputSerializer, ReviewInputSerializer, OrderOutputSerializer, OrderInputSerializer, \
+    OrderOutputSerializer, OrderInputSerializer, \
     OrderLineOutputSerializer, OrderLineInputSerializer, CartOutputSerializer, CartInputSerializer, \
     CartLineOutputSerializer, CartLineInputSerializer
